+++ image/service/postfix/startup.sh
@@ -15,8 +15,8 @@ if [ ! -e "$FIRST_START_DONE" ]; then
   sed -i "s|{{ HOSTNAME }}|${HOSTNAME}|g" ${CONTAINER_SERVICE_DIR}/postfix/assets/config/main.cf
 
   # generate a certificate and key if files don't exists
-  # https://github.com/osixia/docker-light-baseimage/blob/stable/image/service-available/:cfssl/assets/tool/cfssl-helper
-  cfssl-helper ${MMC_MAIL_CFSSL_PREFIX} "${CONTAINER_SERVICE_DIR}/postfix/assets/certs/$MMC_MAIL_SSL_CRT_FILENAME" "${CONTAINER_SERVICE_DIR}/postfix/assets/certs/$MMC_MAIL_SSL_KEY_FILENAME" "${CONTAINER_SERVICE_DIR}/postfix/assets/certs/$MMC_MAIL_SSL_CA_CRT_FILENAME"
+  # https://github.com/osixia/docker-light-baseimage/blob/stable/image/service-available/:ssl-tools/assets/tool/ssl-helper
+  ssl-helper ${MMC_MAIL_SSL_HELPER_PREFIX} "${CONTAINER_SERVICE_DIR}/postfix/assets/certs/$MMC_MAIL_SSL_CRT_FILENAME" "${CONTAINER_SERVICE_DIR}/postfix/assets/certs/$MMC_MAIL_SSL_KEY_FILENAME" "${CONTAINER_SERVICE_DIR}/postfix/assets/certs/$MMC_MAIL_SSL_CA_CRT_FILENAME"
 
   MMC_MAIL_SSL_CA_CRT_PATH="${CONTAINER_SERVICE_DIR}/postfix/assets/certs/${MMC_MAIL_SSL_CA_CRT_FILENAME}"
   MMC_MAIL_SSL_CRT_PATH="${CONTAINER_SERVICE_DIR}/postfix/assets/certs/${MMC_MAIL_SSL_CRT_FILENAME}"