+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 0.1.8
   - Fix opendkim config
-  - Upgrade baseimage: light-baseimage:0.2.4
+  - Upgrade baseimage: light-baseimage:0.2.5
   - Add tls certificates for replication only
 
 ## 0.1.7